+++ lib/src/components/country_picker_pro2/src/config/country_picker_search_config.dart
@@ -1,0 +1,25 @@
+import 'package:flutter/material.dart';
+
+class CountryPickerSearchConfig {
+  final Color? backgroundColor;
+  final Color? textColor;
+  final Color? borderColor;
+  final double? borderWidth;
+  final Color? hintColor;
+  final Icon? icon;
+  final Color? outerBackgroundColor;
+  final String? hintText;
+  final Color? cursorColor;
+
+  const CountryPickerSearchConfig({
+    this.backgroundColor,
+    this.textColor,
+    this.borderColor,
+    this.borderWidth,
+    this.hintColor,
+    this.icon,
+    this.outerBackgroundColor,
+    this.hintText,
+    this.cursorColor,
+  });
+}